Extract sc-executor-common crate
