+++ client/executor/Cargo.toml
@@ -19,6 +19,7 @@ lazy_static = "1.4.0"
 sp-wasm-interface = { version = "2.0.0", path = "../../primitives/wasm-interface" }
 sp-runtime-interface = { version = "2.0.0", path = "../../primitives/runtime-interface" }
 sp-externalities = { version = "2.0.0", path = "../../primitives/externalities" }
+sc-executor-common = { version = "2.0.0", path = "common" }
 parking_lot = "0.9.0"
 log = "0.4.8"
 libsecp256k1 = "0.3.2"
