+++ client/executor/common/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "sc-executor-common"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+
+[dependencies]
+log = "0.4.8"
+derive_more = "0.99.2"
+codec = { package = "parity-scale-codec", version = "1.0.0" }
+wasmi = "0.6.2" # TODO: Remove
+sp-core = { version = "2.0.0", path = "../../../primitives/core" }
+sp-wasm-interface = { version = "2.0.0", path = "../../../primitives/wasm-interface" }
+sp-runtime-interface = { version = "2.0.0", path = "../../../primitives/runtime-interface" }
+sp-serializer = { version = "2.0.0", path = "../../../primitives/serializer" }
