+++ client/executor/src/wasm_runtime.rs
@@ -23,31 +23,13 @@ use crate::{wasmi_execution, error::{Error, WasmError}};
 #[cfg(feature = "wasmtime")]
 use crate::wasmtime;
 use log::{trace, warn};
-
 use codec::Decode;
-
 use sp_core::{storage::well_known_keys, traits::Externalities};
-
 use sp_version::RuntimeVersion;
 use std::{collections::hash_map::{Entry, HashMap}, panic::AssertUnwindSafe};
+use sc_executor_common::wasm_runtime::WasmRuntime;
 
 use sp_wasm_interface::Function;
-
-/// The Substrate Wasm runtime.
-pub trait WasmRuntime {
-	/// Attempt to update the number of heap pages available during execution.
-	///
-	/// Returns false if the update cannot be applied. The function is guaranteed to return true if
-	/// the heap pages would not change from its current value.
-	fn update_heap_pages(&mut self, heap_pages: u64) -> bool;
-
-	/// Return the host functions that are registered for this Wasm runtime.
-	fn host_functions(&self) -> &[&'static dyn Function];
-
-	/// Call a method in the Substrate runtime by name. Returns the encoded result on success.
-	fn call(&mut self, ext: &mut dyn Externalities, method: &str, data: &[u8])
-		-> Result<Vec<u8>, Error>;
-}
 
 /// Specification of different methods of executing the runtime Wasm code.
 #[derive(Debug, PartialEq, Eq, Hash, Copy, Clone)]
